backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pillow==10.1.0
opencv-python==4.8.1.78
ultralytics==8.0.228
scikit-learn==1.3.2
numpy==1.24.3
python-dateutil==2.8.2
